Move the default definitions from test_manifold to a small module, so we avoid double includes.
