+++ test/bases.jl
@@ -4,153 +4,9 @@ using ManifoldsBase: DefaultManifold, ℝ, ℂ, RealNumbers, ComplexNumbers
 using ManifoldsBase: CotangentSpaceType, TangentSpaceType
 using ManifoldsBase: FVector
 using Test
-import Base: +, -, *, copyto!, isapprox
-import ManifoldsBase:
-    allocate,
-    get_vector_orthonormal!,
-    get_coordinates_orthonormal!,
-    get_basis_orthogonal,
-    get_basis_orthonormal
 
-
-struct ProjManifold <: AbstractManifold{ℝ} end
-
-ManifoldsBase.inner(::ProjManifold, p, X, Y) = dot(X, Y)
-ManifoldsBase.project!(::ProjManifold, Y, p, X) = (Y .= X .- dot(p, X) .* p)
-ManifoldsBase.representation_size(::ProjManifold) = (2, 3)
-ManifoldsBase.manifold_dimension(::ProjManifold) = 5
-ManifoldsBase.get_vector_orthonormal(::ProjManifold, p, X, N) = reverse(X)
-
-struct ProjectionTestManifold <: AbstractManifold{ℝ} end
-
-ManifoldsBase.inner(::ProjectionTestManifold, ::Any, X, Y) = dot(X, Y)
-function ManifoldsBase.project!(::ProjectionTestManifold, Y, p, X)
-    Y .= X .- dot(p, X) .* p
-    Y[end] = 0
-    return Y
-end
-ManifoldsBase.manifold_dimension(::ProjectionTestManifold) = 100
-
-struct NonManifold <: AbstractManifold{ℝ} end
-struct NonBasis <: ManifoldsBase.AbstractBasis{ℝ,TangentSpaceType} end
-
-struct NonBroadcastBasisThing{T}
-    v::T
-end
-
-+(a::NonBroadcastBasisThing, b::NonBroadcastBasisThing) = NonBroadcastBasisThing(a.v + b.v)
-*(α, a::NonBroadcastBasisThing) = NonBroadcastBasisThing(α * a.v)
--(a::NonBroadcastBasisThing, b::NonBroadcastBasisThing) = NonBroadcastBasisThing(a.v - b.v)
-
-isapprox(a::NonBroadcastBasisThing, b::NonBroadcastBasisThing) = isapprox(a.v, b.v)
-
-function ManifoldsBase.number_eltype(a::NonBroadcastBasisThing)
-    return typeof(reduce(+, one(number_eltype(eti)) for eti in a.v))
-end
-
-allocate(a::NonBroadcastBasisThing) = NonBroadcastBasisThing(allocate(a.v))
-function allocate(a::NonBroadcastBasisThing, ::Type{T}) where {T}
-    return NonBroadcastBasisThing(allocate(a.v, T))
-end
-allocate(::NonBroadcastBasisThing, ::Type{T}, s::Integer) where {T} = Vector{T}(undef, s)
-
-function copyto!(a::NonBroadcastBasisThing, b::NonBroadcastBasisThing)
-    copyto!(a.v, b.v)
-    return a
-end
-
-function ManifoldsBase.log!(
-    ::DefaultManifold,
-    v::NonBroadcastBasisThing,
-    x::NonBroadcastBasisThing,
-    y::NonBroadcastBasisThing,
-)
-    return copyto!(v, y - x)
-end
-
-function ManifoldsBase.exp!(
-    ::DefaultManifold,
-    y::NonBroadcastBasisThing,
-    x::NonBroadcastBasisThing,
-    v::NonBroadcastBasisThing,
-)
-    return copyto!(y, x + v)
-end
-
-function ManifoldsBase.get_basis_orthonormal(
-    ::DefaultManifold,
-    p::NonBroadcastBasisThing,
-    𝔽::RealNumbers,
-)
-    return CachedBasis(
-        DefaultOrthonormalBasis(𝔽),
-        [
-            NonBroadcastBasisThing(ManifoldsBase._euclidean_basis_vector(p.v, i)) for
-            i in eachindex(p.v)
-        ],
-    )
-end
-function ManifoldsBase.get_basis_orthogonal(
-    ::DefaultManifold,
-    p::NonBroadcastBasisThing,
-    𝔽::RealNumbers,
-)
-    return CachedBasis(
-        DefaultOrthogonalBasis(𝔽),
-        [
-            NonBroadcastBasisThing(ManifoldsBase._euclidean_basis_vector(p.v, i)) for
-            i in eachindex(p.v)
-        ],
-    )
-end
-function ManifoldsBase.get_basis_default(
-    ::DefaultManifold,
-    p::NonBroadcastBasisThing,
-    N::ManifoldsBase.RealNumbers,
-)
-    return CachedBasis(
-        DefaultBasis(N),
-        [
-            NonBroadcastBasisThing(ManifoldsBase._euclidean_basis_vector(p.v, i)) for
-            i in eachindex(p.v)
-        ],
-    )
-end
-
-function ManifoldsBase.get_coordinates_orthonormal!(
-    M::DefaultManifold,
-    Y,
-    ::NonBroadcastBasisThing,
-    X::NonBroadcastBasisThing,
-    ::RealNumbers,
-)
-    copyto!(Y, reshape(X.v, manifold_dimension(M)))
-    return Y
-end
-
-function ManifoldsBase.get_vector_orthonormal!(
-    M::DefaultManifold,
-    Y::NonBroadcastBasisThing,
-    ::NonBroadcastBasisThing,
-    X,
-    ::RealNumbers,
-)
-    copyto!(Y.v, reshape(X, representation_size(M)))
-    return Y
-end
-
-function ManifoldsBase.inner(
-    ::DefaultManifold,
-    ::NonBroadcastBasisThing,
-    X::NonBroadcastBasisThing,
-    Y::NonBroadcastBasisThing,
-)
-    return dot(X.v, Y.v)
-end
-
-ManifoldsBase._get_vector_cache_broadcast(::NonBroadcastBasisThing) = Val(false)
-
-DiagonalizingBasisProxy() = DiagonalizingOrthonormalBasis([1.0, 0.0, 0.0])
+push!(LOAD_PATH, pwd())
+using ManifoldsBaseTestUtils
 
 @testset "Bases" begin
     @testset "Projected and arbitrary orthonormal basis" begin
