+++ test/default_manifold.jl
@@ -1,22 +1,5 @@
 using ManifoldsBase
-using ManifoldsBase:
-    @manifold_element_forwards, @manifold_vector_forwards, @default_manifold_fallbacks
-using ManifoldsBase: DefaultManifold, AbstractNumbers, RealNumbers, ComplexNumbers
-import ManifoldsBase:
-    allocate_result_type,
-    check_point,
-    distance,
-    embed!,
-    exp!,
-    inner,
-    isapprox,
-    log!,
-    number_eltype,
-    parallel_transport_to!,
-    retract!,
-    inverse_retract!
 
-import Base: angle, convert
 using DoubleFloats
 using ForwardDiff
 using LinearAlgebra
@@ -25,192 +8,8 @@ using ReverseDiff
 using StaticArrays
 using Test
 
-struct CustomDefinedRetraction <: ManifoldsBase.AbstractRetractionMethod end
-struct CustomUndefinedRetraction <: ManifoldsBase.AbstractRetractionMethod end
-struct CustomDefinedKeywordRetraction <: ManifoldsBase.AbstractRetractionMethod end
-struct CustomDefinedKeywordInverseRetraction <:
-       ManifoldsBase.AbstractInverseRetractionMethod end
-struct CustomDefinedInverseRetraction <: ManifoldsBase.AbstractInverseRetractionMethod end
-
-struct DefaultPoint{T} <: AbstractManifoldPoint
-    value::T
-end
-DefaultPoint(v::T) where {T} = DefaultPoint{T}(v)
-convert(::Type{DefaultPoint{T}}, v::T) where {T} = DefaultPoint(v)
-Base.size(p::DefaultPoint) = size(p.value)
-Base.eltype(p::DefaultPoint) = eltype(p.value)
-Base.length(p::DefaultPoint) = length(p.value)
-function Base.copyto!(q::DefaultPoint, p::DefaultPoint)
-    copyto!(q.value, p.value)
-    return q
-end
-struct DefaultTVector{T} <: TVector
-    value::T
-end
-DefaultTVector(v::T) where {T} = DefaultTVector{T}(v)
-convert(::Type{DefaultTVector{T}}, ::DefaultPoint, v::T) where {T} = DefaultTVector(v)
-Base.size(X::DefaultTVector) = size(X.value)
-Base.eltype(X::DefaultTVector) = eltype(X.value)
-function Base.fill!(X::DefaultTVector, x)
-    fill!(X.value, x)
-    return X
-end
-function ManifoldsBase.allocate_result_type(
-    ::DefaultManifold,
-    ::typeof(log),
-    ::Tuple{DefaultPoint,DefaultPoint},
-)
-    return DefaultTVector
-end
-function ManifoldsBase.allocate_result_type(
-    ::DefaultManifold,
-    ::typeof(inverse_retract),
-    ::Tuple{DefaultPoint,DefaultPoint},
-)
-    return DefaultTVector
-end
-
-ManifoldsBase.@manifold_element_forwards DefaultPoint value
-ManifoldsBase.@manifold_vector_forwards DefaultTVector value
-ManifoldsBase.@default_manifold_fallbacks ManifoldsBase.DefaultManifold DefaultPoint DefaultTVector value value
-
-function ManifoldsBase._injectivity_radius(::DefaultManifold, ::CustomDefinedRetraction)
-    return 10.0
-end
-function ManifoldsBase._retract!(
-    M::DefaultManifold,
-    q,
-    p,
-    X,
-    t::Number,
-    ::CustomDefinedKeywordRetraction;
-    kwargs...,
-)
-    return retract_custom_kw!(M, q, p, X, t; kwargs...)
-end
-function retract_custom_kw!(
-    ::DefaultManifold,
-    q::DefaultPoint,
-    p::DefaultPoint,
-    X::DefaultTVector,
-    t::Number;
-    scale = 2.0,
-)
-    q.value .= scale .* p.value .+ t .* X.value
-    return q
-end
-function ManifoldsBase._inverse_retract!(
-    M::DefaultManifold,
-    X,
-    p,
-    q,
-    ::CustomDefinedKeywordInverseRetraction;
-    kwargs...,
-)
-    return inverse_retract_custom_kw!(M, X, p, q; kwargs...)
-end
-function inverse_retract_custom_kw!(
-    ::DefaultManifold,
-    X::DefaultTVector,
-    p::DefaultPoint,
-    q::DefaultPoint;
-    scale = 2.0,
-)
-    X.value .= q.value - scale * p.value
-    return X
-end
-
-function ManifoldsBase.retract!(
-    ::DefaultManifold,
-    q::DefaultPoint,
-    p::DefaultPoint,
-    X::DefaultTVector,
-    t::Number,
-    ::CustomDefinedRetraction,
-)
-    q.value .= 2 .* p.value .+ t * X.value
-    return q
-end
-
-function ManifoldsBase.inverse_retract!(
-    ::DefaultManifold,
-    X::DefaultTVector,
-    p::DefaultPoint,
-    q::DefaultPoint,
-    ::CustomDefinedInverseRetraction,
-)
-    X.value .= q.value .- 2 .* p.value
-    return X
-end
-
-struct MatrixVectorTransport{T} <: AbstractVector{T}
-    m::Matrix{T}
-end
-# dummy retractions, inverse retracions for fallback tests - mutating should be enough
-ManifoldsBase.retract_polar!(::DefaultManifold, q, p, X, t::Number) = (q .= p .+ t .* X)
-ManifoldsBase.retract_project!(::DefaultManifold, q, p, X, t::Number) = (q .= p .+ t .* X)
-ManifoldsBase.retract_qr!(::DefaultManifold, q, p, X, t::Number) = (q .= p .+ t .* X)
-function ManifoldsBase.retract_exp_ode!(
-    ::DefaultManifold,
-    q,
-    p,
-    X,
-    t::Number,
-    m::AbstractRetractionMethod,
-    B::ManifoldsBase.AbstractBasis,
-)
-    return (q .= p .+ t .* X)
-end
-
-function ManifoldsBase.retract_pade!(
-    ::DefaultManifold,
-    q,
-    p,
-    X,
-    t::Number,
-    m::PadeRetraction,
-)
-    return (q .= p .+ t .* X)
-end
-function ManifoldsBase.retract_sasaki!(
-    ::DefaultManifold,
-    q,
-    p,
-    X,
-    t::Number,
-    ::SasakiRetraction,
-)
-    return (q .= p .+ t .* X)
-end
-ManifoldsBase.retract_softmax!(::DefaultManifold, q, p, X, t::Number) = (q .= p .+ t .* X)
-ManifoldsBase.get_embedding(M::DefaultManifold) = M # dummy embedding
-ManifoldsBase.inverse_retract_polar!(::DefaultManifold, Y, p, q) = (Y .= q .- p)
-ManifoldsBase.inverse_retract_project!(::DefaultManifold, Y, p, q) = (Y .= q .- p)
-ManifoldsBase.inverse_retract_qr!(::DefaultManifold, Y, p, q) = (Y .= q .- p)
-ManifoldsBase.inverse_retract_softmax!(::DefaultManifold, Y, p, q) = (Y .= q .- p)
-function ManifoldsBase.inverse_retract_nlsolve!(
-    ::DefaultManifold,
-    Y,
-    p,
-    q,
-    m::NLSolveInverseRetraction,
-)
-    return (Y .= q .- p)
-end
-function ManifoldsBase.vector_transport_along_project!(
-    ::DefaultManifold,
-    Y,
-    p,
-    X,
-    c::AbstractVector,
-)
-    return (Y .= X)
-end
-
-
-Base.getindex(x::MatrixVectorTransport, i) = x.m[:, i]
-
-Base.size(x::MatrixVectorTransport) = (size(x.m, 2),)
+push!(LOAD_PATH, pwd())
+using ManifoldsBaseTestUtils
 
 @testset "Testing Default (Euclidean)" begin
     M = ManifoldsBase.DefaultManifold(3)
