+++ test/power.jl
@@ -6,7 +6,8 @@ using StaticArrays
 using LinearAlgebra
 using Random
 
-include("test_manifolds.jl")
+push!(LOAD_PATH, pwd())
+using ManifoldsBaseTestUtils
 
 struct TestVectorSpaceType <: VectorSpaceType end
 
@@ -316,7 +317,7 @@ struct TestArrayRepresentation <: AbstractPowerRepresentation end
     @testset "metric conversion" begin
         M = TestSPD(3)
         N = PowerManifold(M, NestedPowerRepresentation(), 2)
-        e = EuclideanMetric()
+        e = ManifoldsBase.EuclideanMetric()
         p = [1.0 0.0 0.0; 0.0 1.0 0.0; 0.0 0.0 1]
         q = [2.0 0.0 0.0; 0.0 2.0 0.0; 0.0 0.0 1]
         P = [p, q]
